+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Added `\Wizaplace\SDK\Pim\Product\ProductService::updateProduct` and associated classes
 - Added `\Wizaplace\SDK\Vendor\Order\OrderService::getOrderById` and associated classes
 - Added `\Wizaplace\SDK\Vendor\Order\OrderService::listOrders` and associated classes
+- Added `\Wizaplace\SDK\Pim\Product\Attribute\AttributeService::getProductAttributes` and associated classes
+- Added `\Wizaplace\SDK\Pim\Product\Attribute\AttributeService::getProductAttribute` and associated classes
 
 ### Bugfixes